Move queue_chunker.py up a level
