+++ superduperdb/db/mongodb/cdc/handler.py
@@ -6,7 +6,7 @@ from superduperdb import logging
 from superduperdb.container.job import FunctionJob
 from superduperdb.container.task_workflow import TaskWorkflow
 from superduperdb.db.base.db import DB
-from superduperdb.misc.runnable.queue_chunker import QueueChunker
+from superduperdb.misc.queue_chunker import QueueChunker
 from superduperdb.misc.runnable.runnable import Event
 
 from .base import Packet
